pyproject.toml
[project]
name = "musicboy-redux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "python-dotenv>=1.2.1",
    "yt-dlp>=2025.12.8",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
]
